requirements.txt
aurora-data-api==0.2.5
boto3==1.16.10
botocore==1.19.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
fastapi==0.61.1
h11==0.9.0
idna==2.10
jmespath==0.10.0
mysql==0.0.2
mysql-connector-python==8.0.22
mysqlclient==2.0.1
oauthlib==3.1.0
protobuf==3.14.0
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
pyxero==0.9.2
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.8
websockets==8.1
Werkzeug==1.0.1

redis~=3.5.3